refactor(hooks): extract useOperationRegenerations from use-ai-focus
Pull the three useSyncExternalStore subscriptions to the global
operationsManager (topic/challenge/goal regenerations) into a
dedicated hook. The seam between focus-fetch lifecycle and
cross-component visibility is now testable in isolation.

use-ai-focus drops from 558 -> 541 LOC; baseline tightened.
